Code Structure Chenegs , Implemented Interfaces
