+++ apps/frontend/src/types/deployment.ts
@@ -1,7 +1,1 @@
-export interface Deployment {
-  id: string;
-  projectId: string;
-  requestedBy: string;
-  status: "PENDING" | "APPROVED" | "REJECTED";
-  timestamp: string;
-}
+export type { Deployment } from "types";
