+++ apps/frontend/src/types/project.ts
@@ -1,7 +1,1 @@
-export interface Project {
-  id: string;
-  name: string;
-  owner: string;
-  pipelineHash: string;
-  status: "ACTIVE" | "INACTIVE";
-}
+export type { Project } from "types";
