+++ packages/types/auth.ts
@@ -1,0 +1,7 @@
+export interface AuthField {
+    label: string;
+    type: string;
+    id: string;
+    placeholder: string;
+    required: boolean;
+}
