+++ packages/types/deployment.ts
@@ -1,0 +1,7 @@
+export interface Deployment {
+    id: string;
+    projectId: string;
+    requestedBy: string;
+    status: "PENDING" | "APPROVED" | "REJECTED";
+    timestamp: string;
+}
