+++ packages/types/index.d.ts
@@ -1,5 +1,9 @@
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-}
+export * from './user';
+export * from './audit';
+export * from './deployment';
+export * from './project';
+export * from './pipeline';
+export * from './landing';
+export * from './auth';
+export * from './wallet';
+export * from './dashboard';
