+++ packages/types/pipeline.ts
@@ -1,0 +1,7 @@
+export interface PipelineStage {
+    id: number;
+    label: string;
+    time: string;
+    status: 'pending' | 'active' | 'done';
+    icon: string;
+}
