+++ packages/types/project.ts
@@ -1,0 +1,7 @@
+export interface Project {
+    id: string;
+    name: string;
+    owner: string;
+    pipelineHash: string;
+    status: "ACTIVE" | "INACTIVE";
+}
