+++ packages/types/user.ts
@@ -1,0 +1,5 @@
+export interface User {
+    id: string;
+    name: string;
+    email: string;
+}
