+++ packages/types/wallet.ts
@@ -1,0 +1,6 @@
+export interface WalletOption {
+    name: string;
+    description: string;
+    icon: string;
+    iconColor: string;
+}
